+++ modules/nickserv/sendpass.c
@@ -165,7 +165,7 @@ ns_cmd_sendpass(struct sourceinfo *si, int parc, char *parv[])
 		}
 		logcommand(si, CMDLOG_ADMIN, "SENDPASS: \2%s\2 (change key)", name);
 
-		key = random_string(16);
+		key = random_string(config_options.default_pass_length);
 
 		const char *const hash = crypt_password(key);